SafeRouteV1-main/SafeRoute_Backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.9.2
python-dotenv==1.0.0
cachetools==7.0.5
networkx==3.2.1
numpy==1.26.3
scipy==1.16.3
SQLAlchemy==2.0.25
APScheduler==3.11.2
websockets==16.0
zeroconf==0.148.0
osmnx==2.1.0
pandas==3.0.1
geopandas==1.1.3
geopy==2.4.1
pyogrio==0.12.1
pyproj==3.7.2
shapely==2.1.2
scikit-learn==1.8.0
requests==2.32.5
httpx==0.26.0
pytest>=8.0
